Added Details.cshtml and New.cshtml (creating new route and adding it to the database). Removed BikeRouteEntity.cs (Entity Framework uses now BikeRoute.cs (Model layer)

+++ SiBRute/SiBRute.Model.Common/IBikeRoute.cs
@@ -34,7 +34,7 @@ namespace SiBRute.Model.Common
         /// <summary>
         /// Gets or sets the distance of the route
         /// </summary>
-        double Distance { get; set; }
+        int Distance { get; set; }
 
         /// <summary>
         /// Gets or sets the date of the route creation
